refactor: vec2와 pos 추가 정리
