+++ bot/board.cpp
@@ -1,0 +1,1 @@
+#include "board.hpp"
