+++ bot/vec2.tpp
@@ -3,6 +3,11 @@
 
 #include "vec2.hpp"
 
+template <typename T>
+inline vec2<T> vec2<T>::random(T max_y, T max_x) {
+  return vec2<T>(util::randRange(max_y), util::randRange(max_x));
+}
+
 // 연산자 오버로딩
 template <typename T>
 inline bool vec2<T>::operator==(const vec2 &other) const {
@@ -16,7 +21,7 @@ inline bool vec2<T>::operator!=(const vec2 &other) const {
 
 template <typename T>
 inline vec2<T> vec2<T>::operator+(const vec2 &other) const {
-  return from(y + other.y, x + other.x);
+  return vec2(y + other.y, x + other.x);
 }
 
 template <typename T>
@@ -37,14 +42,24 @@ inline vec2<T> &vec2<T>::operator-=(const vec2 &other) {
 }
 
 template <typename T>
-inline vec2<T> vec2<T>::from(T y, T x) {
-  vec2 tmp = {y, x};
-  return tmp;
+inline vec2<T> vec2<T>::operator*(T scalar) const {
+  return from(y * scalar, x * scalar);
 }
 
 template <typename T>
-inline vec2<T> vec2<T>::random(T max_y, T max_x) {
-  return from(util::randRange(max_y), util::randRange(max_x));
+inline vec2<T> &vec2<T>::operator*=(T scalar) {
+  *this = *this * scalar;
+  return *this;
 }
 
+template <typename T>
+inline vec2<T> vec2<T>::operator/(T scalar) const {
+  return from(y / scalar, x / scalar);
+}
+
+template <typename T>
+inline vec2<T> &vec2<T>::operator/=(T scalar) {
+  *this = *this / scalar;
+  return *this;
+}
 #endif  // VEC2_TPP
